requirements.txt
bcrypt==3.1.6
cffi==1.12.2
cycler==0.10.0
Cython==0.29.7
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
django-guardian==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
kiwisolver==1.1.0
Markdown==3.0.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
matplotlib==2.2.0
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
numpy==1.16.2
package-name==0.1
passlib==1.7.1
protobuf==3.7.0
psycopg2-binary==2.8.2
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.0
pystan==2.19.0.0
python-dateutil==2.8.0
pytz==2018.9
rules==2.0.1
six==1.12.0
SQLAlchemy==1.3.2
